Refactor get_actual_signers to utils
