+++ program_admin/util.py
@@ -1,6 +1,7 @@
 from typing import Dict, List
 
 from solana.blockhash import Blockhash
+from solana.keypair import Keypair
 from solana.publickey import PublicKey
 from solana.rpc.async_api import AsyncClient
 from solana.rpc.commitment import Commitment
@@ -107,3 +108,25 @@ def apply_overrides(
         else:
             overridden_permissions[key] = value
     return overridden_permissions
+
+
+def get_actual_signers(
+    signers: List[Keypair], transaction: Transaction
+) -> List[Keypair]:
+    """
+    Given a list of keypairs and a transaction, returns the keypairs that actually need to sign the transaction,
+    i.e. those whose pubkey appears in at least one of the instructions as a signer.
+    """
+    actual_signers = []
+    for signer in signers:
+        instruction_has_signer = [
+            any(
+                signer.public_key == account.pubkey and account.is_signer
+                for account in instruction.keys
+            )
+            for instruction in transaction.instructions
+        ]
+        if any(instruction_has_signer):
+            actual_signers.append(signer)
+
+    return actual_signers
